+++ evm/src/interfaces/IMatchingEngineState.sol
@@ -15,10 +15,10 @@ interface IMatchingEngineState {
      * @return penalty The penalty amount.
      * @return userReward The user reward amount.
      */
-    function calculateDynamicPenalty(uint128 amount, uint128 blocksElapsed)
+    function calculateDynamicPenalty(uint64 amount, uint64 blocksElapsed)
         external
         view
-        returns (uint128 penalty, uint128 userReward);
+        returns (uint64 penalty, uint64 userReward);
 
     /**
      * @notice Calculates the dynamic penalty for the specified auction.
@@ -29,7 +29,7 @@ interface IMatchingEngineState {
     function calculateDynamicPenalty(bytes32 auctionId)
         external
         view
-        returns (uint128 penalty, uint128 userReward);
+        returns (uint64 penalty, uint64 userReward);
 
     /**
      * @notice Returns the original `deployer` of the contracts.
@@ -82,12 +82,12 @@ interface IMatchingEngineState {
     /**
      * @notice Returns the transfer amount for a particular auction.
      */
-    function getAuctionAmount(bytes32 auctionId) external view returns (uint128);
+    function getAuctionAmount(bytes32 auctionId) external view returns (uint64);
 
     /**
      * @notice Returns the security deposit for a particular auction.
      */
-    function getSecurityDeposit(bytes32 auctionId) external view returns (uint128);
+    function getSecurityDeposit(bytes32 auctionId) external view returns (uint64);
 
     /**
      * @notice Returns the status of the specified `auctionId`.
@@ -125,7 +125,7 @@ interface IMatchingEngineState {
     /**
      * @notice Returns the number of blocks elapsed since the auction started.
      */
-    function getAuctionBlocksElapsed(bytes32 auctionId) external view returns (uint128);
+    function getAuctionBlocksElapsed(bytes32 auctionId) external view returns (uint64);
 
     /**
      * @notice Returns a boolean which indicates whether the specified `FastFill` has been redeemed.
