+++ evm/src/interfaces/IMatchingEngineTypes.sol
@@ -35,13 +35,13 @@ struct LiveAuctionData {
     // The initial bidder of the auction.
     address initialBidder;
     // The block number at which the auction started.
-    uint128 startBlock;
+    uint64 startBlock;
     // The amount of tokens to be sent to the user.
-    uint128 amount;
+    uint64 amount;
     // The additional deposit made by the highest bidder.
-    uint128 securityDeposit;
+    uint64 securityDeposit;
     // The bid price of the highest bidder.
-    uint128 bidPrice;
+    uint64 bidPrice;
 }
 
 struct LiveAuctionInfo {
